requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alpaca-py==0.30.1
alpaca-trade-api==3.2.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
deprecation==2.1.0
dnspython==1.16.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.0
flatbuffers==24.3.25
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gnews==0.3.7
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
html5lib==1.1
huggingface-hub==0.24.6
idna==3.8
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
keras==2.15.0
libclang==18.1.1
lxml==5.3.0
lxml_html_clean==0.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
namex==0.0.8
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.2
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
readability-lxml==0.8.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.24.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sseclient-py==1.8.0
statsmodels==0.14.2
sympy==1.13.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.4
wrapt==1.14.1
wsproto==1.2.0
yarl==1.11.1
yfinance==0.2.43